refactor text

+++ src/render/info.hpp
@@ -12,15 +12,10 @@ struct Info : base {
 	void init() override {};
 	void render() override {
 		std::string counter = std::to_string(d->window.global.serial);
-		d->text.rMono32(counter,
-			static_cast<int>(d->window.w) - 16,
-			16,
-			Text::Align::RIGHT);
+		d->text.rMono32(counter, d->window.w - 16.0f, 16, Text::Align::RIGHT);
 
-		d->text.rMono32(BUILD_TIMESTAMP,
-			16,
-			static_cast<int>(d->window.h) - 48,
-			Text::Align::LEFT);
+		d->text.rMono32(
+			BUILD_TIMESTAMP, 16, d->window.h - 48.0f, Text::Align::LEFT);
 	};
 };
